Different way of doing echo in shell
Rather than having a dummy stdout producer and echoing directly from the stdin consumer, use the linebuff as an intermediate queue and have the stdout producer use that. I did this for backspace: so that on a backspace I could do a carriage return and re-print the whole line buffer from the start.